+++ src/Model/Holiday.php
@@ -14,9 +14,13 @@ namespace Umulmrum\Holiday\Model;
 use DateTimeImmutable;
 use DateTimeInterface;
 use DateTimeZone;
+use InvalidArgumentException;
 use Umulmrum\Holiday\Constant\HolidayType;
 use Umulmrum\Holiday\Formatter\HolidayFormatterInterface;
 use Umulmrum\Holiday\Provider\DateCreatorTrait;
+
+use function mb_strlen;
+use function preg_match;
 
 class Holiday
 {
@@ -28,7 +32,21 @@ class Holiday
     /** @var string */
     public const CREATE_DATE_FORMAT = '!Y-m-d';
 
-    public function __construct(private string $name, private string $simpleDate, private int $type = HolidayType::OTHER) {}
+    public function __construct(
+        private readonly string $name,
+        private readonly string $simpleDate,
+        private readonly int $type = HolidayType::OTHER,
+    ) {
+        if ($this->name === '') {
+            throw new InvalidArgumentException('Holiday name must not be empty.');
+        }
+        if (mb_strlen($this->simpleDate) !== 10 || preg_match('#^\d{4}-\d{2}-\d{2}$#', $this->simpleDate) !== 1) {
+            throw new InvalidArgumentException('Date must be in format YYYY-MM-DD, got ' . $this->simpleDate);
+        }
+        if ($this->type < 0 || $this->type > HolidayType::ALL) {
+            throw new InvalidArgumentException('Type must be a combination of types as defined in the HolidayTypes class, got ' . $this->type);
+        }
+    }
 
     public static function create(string $name, string $date, int $type = HolidayType::OTHER): self
     {
@@ -37,7 +55,7 @@ class Holiday
 
     public static function createFromDateTime(string $name, DateTimeInterface $date, int $type = HolidayType::OTHER): self
     {
-        return new self($name, $date->format((string) static::DISPLAY_DATE_FORMAT), $type);
+        return new self($name, $date->format(static::DISPLAY_DATE_FORMAT), $type);
     }
 
     public function getName(): string
@@ -53,7 +71,7 @@ class Holiday
     public function getDate(?DateTimeZone $dateTimeZone = null): DateTimeImmutable
     {
         // @phpstan-ignore-next-line
-        return DateTimeImmutable::createFromFormat((string) static::CREATE_DATE_FORMAT, $this->simpleDate, $dateTimeZone);
+        return DateTimeImmutable::createFromFormat(static::CREATE_DATE_FORMAT, $this->simpleDate, $dateTimeZone);
     }
 
     public function getType(): int