+++ src/Provider/Canada/Alberta.php
@@ -1,0 +1,35 @@
+<?php
+
+/*
+ * This file is part of the umulmrum/holiday package.
+ *
+ * (c) Stefan Kruppa
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Umulmrum\Holiday\Provider\Canada;
+
+use Umulmrum\Holiday\Constant\HolidayType;
+use Umulmrum\Holiday\Model\HolidayList;
+use Umulmrum\Holiday\Provider\CommonHolidaysTrait;
+use Umulmrum\Holiday\Provider\Religion\ChristianHolidaysTrait;
+
+class Alberta extends Canada
+{
+    use ChristianHolidaysTrait;
+    use CommonHolidaysTrait;
+
+    public function calculateHolidaysForYear(int $year): HolidayList
+    {
+        $holidays = parent::calculateHolidaysForYear($year);
+        if ($year >= 1990) {
+            $holidays->add($this->getFamilyDay($year));
+        }
+        $holidays->add($this->getVictoriaDay($year, HolidayType::DAY_OFF));
+        $holidays->add($this->getBoxingDay($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF | HolidayType::PARTIAL_ONLY));
+
+        return $holidays;
+    }
+}